Use environment variables
Move backend environment variables out of application.properties and
into docker-compose.yml.

Move jwt.secret to an external file and inject it via
docker-compose.yml and the docker-entrypoint.sh files in auth-service
and booking-service.

Shorten com.booking.cab.frontend.origin to frontend.origin because
it's used both in the booking and auth services.